Moved `ssh_with_commands` and refactored `subprocess` email patterns
`ssh_with_commands` was written for `download_rsmas.py` but the pattern is used by multiple scripts for data download and sending emails. This function is now used instead of `ssh`ing via a subprocess call inline. Note that `ssh_with_commands` does make a subprocess call, but now we have isolated it to a single function.

Future work: Return Exceptions along with the process error code from `ssh_with_commands`. 

@lily-wittle
